backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
aiomysql==0.2.0
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.2.0
pydantic-settings==2.7.1
email-validator==2.2.0
redis==5.2.1
httpx==0.27.2